fix mocks

+++ state-chain/traits/src/mocks/epoch_info.rs
@@ -69,7 +69,7 @@ macro_rules! impl_mock_epoch_info {
 		impl EpochInfo for MockEpochInfo {
 			type ValidatorId = $account_id;
 			type Amount = $balance;
-			type BlockNumber = u32;
+			type BlockNumber = u64;
 
 			fn current_validators() -> Vec<Self::ValidatorId> {
 				CURRENT_VALIDATORS.with(|cell| cell.borrow().clone())
